+++ cmd/deploy.go
@@ -18,12 +18,14 @@ import (
 )
 
 const (
-	VALUES_FLAG                   = "values"
-	CHART_NAME                    = "groundcover"
-	DEFAULT_GROUNDCOVER_RELEASE   = "groundcover"
-	DEFAULT_GROUNDCOVER_NAMESPACE = "groundcover"
-	GROUNDCOVER_URL               = "https://app.groundcover.com"
-	HELM_REPO_URL                 = "https://helm.groundcover.com"
+	VALUES_FLAG                    = "values"
+	CHART_NAME                     = "groundcover"
+	DEFAULT_GROUNDCOVER_RELEASE    = "groundcover"
+	DEFAULT_GROUNDCOVER_NAMESPACE  = "groundcover"
+	COMMIT_HASH_LABEL_NAME_FLAG    = "git-commit-hash-label-name"
+	REPOSITORY_URL_LABEL_NAME_FLAG = "git-repository-url-label-name"
+	GROUNDCOVER_URL                = "https://app.groundcover.com"
+	HELM_REPO_URL                  = "https://helm.groundcover.com"
 )
 
 func init() {
@@ -31,6 +33,12 @@ func init() {
 
 	DeployCmd.PersistentFlags().StringSliceP(VALUES_FLAG, "f", []string{}, "specify values in a YAML file or a URL (can specify multiple)")
 	viper.BindPFlag(VALUES_FLAG, DeployCmd.PersistentFlags().Lookup(VALUES_FLAG))
+
+	DeployCmd.PersistentFlags().String(COMMIT_HASH_LABEL_NAME_FLAG, "", "label name that contains app's git commit hash")
+	viper.BindPFlag(COMMIT_HASH_LABEL_NAME_FLAG, DeployCmd.PersistentFlags().Lookup(COMMIT_HASH_LABEL_NAME_FLAG))
+
+	DeployCmd.PersistentFlags().String(REPOSITORY_URL_LABEL_NAME_FLAG, "", "label name that contains app's git repoistory url")
+	viper.BindPFlag(REPOSITORY_URL_LABEL_NAME_FLAG, DeployCmd.PersistentFlags().Lookup(REPOSITORY_URL_LABEL_NAME_FLAG))
 }
 
 var DeployCmd = &cobra.Command{
@@ -236,6 +244,8 @@ func defaultChartValues(clusterName, apikey string) map[string]interface{} {
 	chartValues := make(map[string]interface{})
 	chartValues["clusterId"] = clusterName
 	chartValues["global"] = map[string]interface{}{"groundcover_token": apikey}
+	chartValues["commitHashLabelName"] = viper.GetString(COMMIT_HASH_LABEL_NAME_FLAG)
+	chartValues["repositoryUrlLabelName"] = viper.GetString(REPOSITORY_URL_LABEL_NAME_FLAG)
 
 	return chartValues
 }